+++ src/parameters.h
@@ -20,41 +20,41 @@
 
 struct Parameters
 {
-	/// Size of the patch (spatial)
-	unsigned k;
-	/// Size of the patch (temporal)
-	unsigned kt;
-	/// Step size
-	unsigned p;
+    /// Size of the patch (spatial)
+    uint64_t k;
+    /// Size of the patch (temporal)
+    uint64_t kt;
+    /// Step size
+    uint64_t p;
 
-	/// Maximum number of similar patches
-	unsigned N;
-	/// Maximum distance threshold
-	float tau;
-	/// Correcting parameter in the distance computation
-	float d;
+    /// Maximum number of similar patches
+    uint64_t N;
+    /// Maximum distance threshold
+    float tau;
+    /// Correcting parameter in the distance computation
+    float d;
 
-	/// Number of frames forward (and backward) used during the search
-	unsigned Nf;
-	/// Size of the search region in the reference frame
-	unsigned Ns;
-	/// Size of the search region in the other frame
-	unsigned Npr;
-	/// Maximum number of matches kept for a frame
-	unsigned Nb;
+    /// Number of frames forward (and backward) used during the search
+    uint64_t Nf;
+    /// Size of the search region in the reference frame
+    uint64_t Ns;
+    /// Size of the search region in the other frame
+    uint64_t Npr;
+    /// Maximum number of matches kept for a frame
+    uint64_t Nb;
 
-	/// Type of the 2D tranform
-	unsigned T_2D;
-	/// Type of the 1D tranform 
-	unsigned T_3D;
-	/// Threshold if it's a hard thresholding step
-	float lambda3D = 0.;
+    /// Type of the 2D tranform
+    uint64_t T_2D;
+    /// Type of the 1D tranform
+    uint64_t T_3D;
+    /// Threshold if it's a hard thresholding step
+    float lambda3D = 0.;
 
-	/// Motion compensation for 3d patches
-	bool mc = false;
+    /// Motion compensation for 3d patches
+    bool mc = false;
 
-	/// Border of the tile when using the multithreading
-	int n = 16;
+    /// Border of the tile when using the multithreading
+    int n = 16;
 };
 
 #endif
